refactor: swr을 이용해서 데이터를 페치하여 useEffect 종속성 제거 (#102)
